requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
httpx==0.24.1
python-dotenv==1.0.1
pydantic==2.8.1
slowapi==0.1.7
tenacity==9.2.2
google-generativeai==0.7.0
